pyproject.toml
[tool.poetry]
name = "xobot"
version = "0.1.0"
description = ""
authors = ["idilettant <dedenev1985@gmail.com>"]
license = "MIT"
packages = [
  { include = "tictactoe" },
]


[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
requests = "^2.25.1"
dataclass-factory = "^2.10.2"
pook = "^1.0.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.11.1"
wemake-python-styleguide = "^0.15.2"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
flake8-cognitive-complexity = "^0.1.0"
pyngrok = "^5.0.5"
mypy = "^0.812"
python-dotenv = "^0.17.1"
pook = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tictactoe = "tictactoe.scripts.xobot:main"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
    "integration",
]
